LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-name-cross-shadow-expected.txt


FAIL Should be able to set anchor-name to a shadow DOM part and anchor to it assert_equals: expected 15 but got 0
PASS Should be able to set anchor-name to the shadow host and anchor to it
